USA Medical Cost Analysis
This is a fairly basic dataset that I have performed EDA and created a linear regression model to predict the charge on the patient based on various variables such as sex, age, bmi etc.

There is fairly extensive commentry in this program to discuss what is being done at each step in the process.